added an http health route. refactored main.go
