+++ gifbot/Gopkg.lock
@@ -8,6 +8,18 @@
   version = "v0.17.0"
 
 [[projects]]
+  name = "github.com/gorilla/context"
+  packages = ["."]
+  revision = "1ea25387ff6f684839d82767c1733ff4d4d15d0a"
+  version = "v1.1"
+
+[[projects]]
+  name = "github.com/gorilla/mux"
+  packages = ["."]
+  revision = "24fca303ac6da784b9e8269f724ddeb0b2eea5e7"
+  version = "v1.5.0"
+
+[[projects]]
   name = "github.com/gorilla/websocket"
   packages = ["."]
   revision = "ea4d1f681babbce9545c9c5f3d5194a789c89f5b"
@@ -22,6 +34,6 @@
 [solve-meta]
   analyzer-name = "dep"
   analyzer-version = 1
-  inputs-digest = "93bb33f2137b28cbe7a26b412b2de8945d8803847dda718499775b08bba59a30"
+  inputs-digest = "d7e2489ae6b95e6b2acff3b2c027f8036a153eb078a19eff5a80506d4fae6097"
   solver-name = "gps-cdcl"
   solver-version = 1
