+++ gifbot/Gopkg.toml
@@ -24,3 +24,7 @@
 [[constraint]]
   name = "github.com/bwmarrin/discordgo"
   version = "0.17.0"
+
+[[constraint]]
+  name = "github.com/gorilla/mux"
+  version = "1.5.0"
